Change uses of DateTimeOffset to just plain DateTime (the extra functionality was not needed.
Change the DaysNeeded variable from a uint to a TimeSpan.

Use a function to decide how the DateCalc variable is populated.

Correct the formula for the final date to be able to start.

Make sure that Reset is greater than 0

Add code to do console prompting if a command line option is not present if command line options are set to be used.

Rename end to AnnounceEnd

Add Announce function to do the printing of the information.

Use result.WithParsed so that the information is only printed if the command line options were parsed succesfully (even if they aren't used).