Add new implementation for the calculation class.
Class instantiation is now done using a couple of factory classes, one abstract and one traditional. This should make the code easier to read (hopefully).

Also, JSON data is now calculated in the class constructor that has the command-line parameters passed to it because that's the only place that JSON will ever be needed.