Make Key.Ask into a generic function and use overrides instead of a long parameter list. This way I won't have to use the different type's parse function (not directly, anyway)
Add Key.AskObj to emit the keyboard input as an object. This feature may not be used in this project, but it may find its way into other projects.

Add structs to store prompts and exception messages. This makes the code more readable and removes a source of inconsistency.

STOFactory is no longer an abstract class, there was really no reason for it to be one in the first place